18371 wrong sigset_t size with ucontext.h first and XPG4v2 on amd64 Reviewed by: Bill Sommerfeld <sommerfeld@hamachi.org> Reviewed by: Toomas Soome <tsoome@me.com> Reviewed by: Peter Tribble <peter.tribble@gmail.com> Approved by: Dan McDonald <danmcd@edgecast.io>

+++ usr/src/uts/intel/sys/ucontext.h
@@ -55,7 +55,7 @@ extern "C" {
 #ifndef	_SIGSET_T
 #define	_SIGSET_T
 typedef	struct {	/* signal set type */
-	unsigned long	__sigbits[4];
+	unsigned int	__sigbits[4];
 } sigset_t;
 #endif /* _SIGSET_T */
